Include Source Code Pro as webfont
I tried using the CDN link with Google Fonts [1], but it doesn't
include all of the line-drawing symbols, so the result involves
pulling those in from other fonts on my system (bad idea).  So, just
including a copy of it here.  Downside is the entire font will be
downloaded, rather than just a subset, but 95k is not enormous.

[1] https://fonts.google.com/specimen/Source+Code+Pro?query=source&selection.family=Source+Code+Pro

Also, adds "monospace" as a fallback for when the font download
fails (e.g. on a really old browser) [2], which is certainly better
than using a proportional font.

[2] http://caniuse.com/#search=woff